Form functionality WIP
I changed the date forms from a select list to an input date type. I styled the CSS to match the rest of the form styling, but I still need to figure out how to match the input date alignment with the other select forms.

In JavaScript, the goal is to log all of the form input values when you click the button "Let's Go Places". 

I created an object that stores all the id's of each of the form's elements and paired them with a value property using dot notation.

Next, I created a function to log the object that is storing all the form data.

It currently logs the correct values, but only for one whole form state change. If you click the button again after changing your form options, it will still log the first state change only. I need to figure out how to make it adapt to a state change anytime the user changes any of the form inputs.