JavaScript, HTML, and CSS changes
JavaScript: Changed my object from a const to let, so that it's property values could be changed on the fly. I was then able to console.log the new object which logs all the input data changed by the user.

HTML:  I added a new class, .content to the flex parent container of the main content. The new .content class is styled with { position: relative; z-index: 1; } with the purpose of enabling the main content to float/scroll above the background image.

Added a new div with a class .myBackground to migrate the background image styling properties out of the body element . The goal was to not have the background image tile or have a blank white empty space when the screen size reduces width and starts scrolling.

CSS: Started to build for responsive design and established a media query for (max-width: 768px), so that any resolution lower than 768 will start stacking the layout responsively.