set up form validation
I got stuck and had some assistance on figuring out how to capture the user input's data from my objForm object upon selection. Once I was able to log all the live input selections by the user with my submitForm() function, I was able to then start figuring out how to set up a validation function with a conditional statement. With assistance I was able to learn how to create a function which can return the properties from my objForm object as an array by using Object.keys(); function and assigned objForm inside of that function as an argument. The validateForm() function then console.logs the length of the array. Next the validateForm() function declares a block scope let statement variable which is let errors = 0;. Then there is a forEach() method used to provide a function for each array element, which uses an if statment that checks if the function inputIsEmpty() in increments of 1, then it logs how many errors out of 5 there are. I had assistance to learn a little bit of jQuery to set up the conditional for the disabled button. If there are errors > 0 then the button will be disabled, else if there are 0 errors, the button is enabled.